* implement recentObjects local storage logic as a class * moves logout logic to the new recentObjects class * update components to use the class
